a1

+++ pybricks/pyBricks1/main.py
@@ -105,7 +105,7 @@ async def water_the_green_plants():
         await WheelController.move_wheels_backward_in_straight_line(float(220))
 
     await WheelController.wheel_right_turn()
-    await WheelController.move_wheels_forward_in_straight_line(float(160))
+    await WheelController.move_wheels_forward_in_straight_line(float(165))
     await WheelController.wheel_left_turn()
     await WheelController.move_wheels_forward_in_straight_line(float(60))
     await WheelController.move_wheels_towards_element_then_stop_at_marker()
